Add some spans to ByteBuffer
Motivation:

Users would like to be able to access the underlying memory of a
ByteBuffer, as evidenced by the plethora of `withUnsafe*` methods
that ByteBuffer has. As Swift 6.2 has introduced some initial
APIs for safe memory access to underlying storage, we should offer
similar APIs on ByteBuffer to enable users to get safer access to
that storage.

For now, the obvious APIs to be able to supplement are:

- withUnsafeReadableBytes
- withUnsafeMutableReadableBytes
- writeWithUnsafeMutableWritableBytes

We can also offer some new APIs to allow initializing a buffer
directly from an OutputSpan.

Note that we can only do this because the Language Steering Group
has pinky promised that they will not break the "Lifetimes"
experimental feature: see
https://forums.swift.org/t/experimental-support-for-lifetime-dependencies-in-swift-6-2-and-beyond/78638
for more details. We are taking them at their word, and so we are
enabling that feature.

Modifications:

Many new methods and tests.

Result:

Safer access.