Tweak test availability to avoid segfaults
Motivation

Availability guards on test functions don't work the way
we'd want them to: in particular, xctest still ends up
calling these functions even on platforms that are less
available than the functions in question. We need to
move these to checks in in the function themselves.

Modifications

Replace @available with guard available

Result

No more segfaults in tests.