Avoid zombie pipe channels when receiving writeEOF
Motivation

In rare cases where the inbound side of a pipe channel is already
closed (either because it was never open or because it was closed
during use), we can get into trouble if the reader drops the
read side of our write pipe. In that context, the Linux kernel
will deliver an EPOLLHUP, but while we'll close the FD we'll let
the channel hang out as a zombie. This is, obviously, suboptimal.

Modifications

In writeEOF, we check whether the read side is already closed. If
it is, we trigger a close internally to shut the channel down.
Added a test for this as needed, and modified an existing test
that would now trip over this behaviour. Also added a new test
for the reverse case to ensure we don't end up with zombies
there (we don't).

Results

No zombie channels.