Generate matrix script finds manifests in subdirs
Motivation:

The genereate matrix script attempts to derive the minimum Swift version
supported by the package so that it doesn't run CI on versions which are
older than that supported. At the moment it can't cope with repositories
with packages in subdirectories like swift-nio-examples.

Modifications:

The script looks for manifests in subdirectories.

Result:

CI should generate more appropriate jobs.