Enable per-message GSO on Linux
Motivation:

NIO has existing GSO support. This GSO support is applied at
Channel scope, which makes it somewhat less than ideal, as it
forces the datagram sizes to be known statically ahead of
time. In real applications this is less common than
being able to do this on a per-superbuffer
context.

Linux supports this already by using sendmsg/sendmmsg
control data. We just need to wire this up in NIO.

An extension to this is that we can also do per-message
GRO in Linux. I'll tackle this in a later patch to keep
the diffs small.

Modifications:

- Adding a new field on AddressedEnvelope Metadata.
- Reject setting this field on non-Linux devices.
- On Linux, use this to set the control data.
- Added new tests that validate the behaviour is correct.

Result:

Enable per-message GSO.